+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-shortener` will be documented in this file.
 
+### 4.0.0
+- Added Google shortener support
+- Can now rotate through various services and various accounts per service
+
 ### 3.0.0
 - Updated for Bitly's changes to API authentication